SegRef3D/requirements.txt
altgraph==0.17.4
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pefile==2023.2.7
pillow==11.3.0
pydicom==3.0.1
pyinstaller-hooks-contrib==2025.6
pyqt6==6.9.1
pyqt6-qt6==6.9.1
pyqt6-sip==13.10.2
pywin32-ctypes==0.2.3
scikit-image==0.22.0
scipy==1.16.0
setuptools==80.9.0
shapely==2.1.1
svgpathtools==1.7.1
svgwrite==1.4.3
torch==2.3.0
trimesh==4.7.1
